+++ mender-update/main.cpp
@@ -12,27 +12,21 @@
 //    See the License for the specific language governing permissions and
 //    limitations under the License.
 
-#include <iostream>
 #include <memory>
 
 using namespace std;
 
-#include <common/json/json.hpp>
-#include <common/kv_db/kv_db.hpp>
+#include <common/json.hpp>
+#include <common/kv_db.hpp>
 
-// This function only knows about the interfaces to JSON and KeyValueDB
 void hello_world(std::shared_ptr<json::Json> j, std::shared_ptr<kv_db::KeyValueDB> db) {
 	j->hello_world();
 	db->hello_world();
 }
 
-#include <common/json/impl/nlohmann/nlohmann_json.hpp>
-#include <common/kv_db/impl/lmdb/lmdb_kv_db.hpp>
-
 int main() {
-	// It is here that we make objects from concrete types
-	shared_ptr<json::NlohmannJson> j = make_shared<json::NlohmannJson>();
-	shared_ptr<kv_db::LMDB_KeyValueDB> db = make_shared<kv_db::LMDB_KeyValueDB>();
+	shared_ptr<json::Json> j = make_shared<json::Json>();
+	shared_ptr<kv_db::KeyValueDB> db = make_shared<kv_db::KeyValueDB>();
 
 	hello_world(j, db);
